requirements.txt
Django>=4.0
djangorestframework
django-cors-headers
djangorestframework-simplejwt
gunicorn
whitenoise
dj-database-url
psycopg2-binary
python-dotenv
